requirements.txt
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain
langchain-google-genai
python-dotenv
requests
